app/requirements.txt
# Telegram Bot
python-telegram-bot==21.6

# OpenAI API client
openai>=1.40.0

# 환경 변수 관리
python-dotenv>=1.0.1

# 스케줄링
apscheduler>=3.10.4

# 한국 시간대, 날짜 처리
pendulum>=3.0.0
pytz>=2025.2

# 고급 정규식
regex>=2025.8.29

# 기타 유틸
tqdm>=4.67.1
